moved and renamed tests
